+++ assets/javascripts/kitten/karl/headers/toggleLink.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export const KarlToggleClassLink = props => {
+  const { label, elementSelector, toggledClass } = props
+  const onClick = event => {}
+
+  return (
+    <a className="karl-link" href="#">
+      {label}
+    </a>
+  )
+}